refactor(tree): разбить код на модули
* Перенёс код класса Card в отдельный файл в директории components, сделал из него модуль, импортировал в index.js
* То же самое с массивом дефолтных карточек. Теперь они в отдельном файле в директории constants, модулем импортирован в index.js
