+++ index.html
@@ -134,7 +134,6 @@
   </div>
 
   <script src="scripts/validate.js"></script>
-  <script src="scripts/initial-cards.js"></script>
-  <script src="scripts/index.js"></script>
+  <script type="module" src="scripts/index.js"></script>
 </body>
 </html>
