--- scripts/initial-cards.js
@@ -1,26 +1,0 @@
-const initialCards = [
-  {
-    name: 'Сундуки',
-    link: './images/sundki.jpeg'
-  },
-  {
-    name: 'Хаыйракан',
-    link: './images/hayirakan.jpeg'
-  },
-  {
-    name: 'Саяно-Шушенская ГЭС',
-    link: './images/sayano-shushenskaya-ges.jpg'
-  },
-  {
-    name: 'Ленские столбы',
-    link: './images/lenskie-stolby.jpeg'
-  },
-  {
-    name: 'Ергаки',
-    link: './images/ergaki.jpg'
-  },
-  {
-    name: 'Курганская пещера',
-    link: './images/kungurskaya-peshera.jpeg'
-  }
-];
